backend/requirements.txt
# -----------------------------
# Core Backend
# -----------------------------
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
sse-starlette
cloudinary

# -----------------------------
# Vector Database
# -----------------------------
qdrant-client

# -----------------------------
# LLM APIs
# -----------------------------
groq
google-generativeai

# -----------------------------
# OCR
# -----------------------------
google-cloud-vision

# -----------------------------
# PDF & Image Processing
# -----------------------------
pymupdf
Pillow
opencv-python

# -----------------------------
# Text Processing
# -----------------------------
langdetect
tiktoken

# -----------------------------
# Embeddings
# -----------------------------
sentence-transformers
FlagEmbedding
